Added slice for trial and reorganized slice position in file hierarchy
