+++ src/components/tracks/IndexBased.jsx
@@ -8,13 +8,13 @@ import { schemeTableau10 } from 'd3';
 
 import { gene } from './gene.js'
 import { panComparison, zoomComparison, moveMiniview, selectMiniviews, updateData, changeMiniviewColor, changeMiniviewVisibility, movePreview, changePreviewVisibility, updatePreview, selectComparison } from 'features/miniview/miniviewSlice.js'
-import { changeZoom, pan, selectBasicTracks, setSelection, clearSelection, updateTrack, addBasicTrack, deleteAllOrthologTracks } from "./basicTrackSlice";
-import { addAnnotation, selectAnnotations, selectSearch, removeAnnotation, addOrtholog, clearOrthologs, selectOrthologs } from "features/annotation/annotationSlice";
+import { changeZoom, pan, selectBasicTracks, setSelection, clearSelection, updateTrack, addBasicTrack, deleteAllOrthologTracks } from "../../redux/slices/basicTrackSlice.js";
+import { addAnnotation, selectAnnotations, selectSearch, removeAnnotation, addOrtholog, clearOrthologs, selectOrthologs } from "redux/slices/annotationSlice.js";
 import { line } from 'd3-shape';
 import Window from "features/miniview/Window.js";
-import { selectDraggables, addDraggable, clearDraggables } from "features/draggable/draggableSlice.js";
+import { selectDraggables, addDraggable, clearDraggables } from "redux/slices/draggableSlice.js";
 import TrackControls from "./TrackControls.jsx";
-import { selectGenome } from "./genomeSlice.js";
+import { selectGenome } from "../../redux/slices/genomeSlice.js";
 // import { getFile } from '../../features/parsers/fetchData';
