+++ src/pages/Comparison.jsx
@@ -6,12 +6,12 @@ import DragContainer from '../features/draggable/DragContainer';
 import AlternateDraggable from '../features/draggable/AlternateDraggable'
 import { useSelector, useDispatch } from 'react-redux';
 import { addComparison, selectMiniviews, clearComparisons, moveCollabPreview } from '../features/miniview/miniviewSlice';
-import { moveAlternateDraggable, selectAlternateDraggables } from '../features/draggable/alternateDraggableSlice';
-import { deleteAllDraggables, selectDraggables, selectGroup, setDraggables } from '../features/draggable/draggableSlice';
+import { moveAlternateDraggable, selectAlternateDraggables } from '../redux/slices/alternateDraggableSlice';
+import { deleteAllDraggables, selectDraggables, selectGroup, setDraggables } from '../redux/slices/draggableSlice';
 import { css } from '@emotion/react';
 import { useState } from 'react';
-import { addDraggable, removeDraggable } from '../features/draggable/draggableSlice';
-import { addAlternateDraggable, removeAlternateDraggable } from '../features/draggable/alternateDraggableSlice';
+import { addDraggable, removeDraggable } from '../redux/slices/draggableSlice';
+import { addAlternateDraggable, removeAlternateDraggable } from '../redux/slices/alternateDraggableSlice';
 import { addMiniview, removeMiniview, selectComparison, removeComparison } from '../features/miniview/miniviewSlice';
 import { Switch, Button, Stack, Divider, FormControl, FormControlLabel, Drawer } from '@mui/material'
 import testing_array2 from '../data/testing_array2';
@@ -19,7 +19,7 @@ import testing_array3 from '../data/testing_array3';
 import { Typography, Slider } from '@mui/material';
 import { CustomDragLayer } from 'features/draggable/CustomDragLayer';
 import BasicTrack from 'components/tracks/BasicTrack';
-import { selectBasicTracks, addComplicatedTrack, removeBasicTrack, deleteAllBasicTracks, updateTrack, toggleTrackType, updateBothTracks, changeBasicTrackColor,} from 'components/tracks/basicTrackSlice';
+import { selectBasicTracks, addComplicatedTrack, removeBasicTrack, deleteAllBasicTracks, updateTrack, toggleTrackType, updateBothTracks, changeBasicTrackColor,} from 'redux/slices/basicTrackSlice';
 // import { pullInfo } from 'features/parsers/gffParser'; 
 import { text } from "d3-request";
 import $ from 'jquery';
@@ -34,8 +34,8 @@ import parseGFF from 'features/parsers/gffParser';
 import { parseSubmittedGFF, parseSubmittedCollinearity } from 'features/parsers/gffParser';
 import _, { reject } from 'lodash';
 import OrthologLinks from 'components/tracks/OrthologLinks';
-import { addAnnotation, clearSearches, addSearch } from 'features/annotation/annotationSlice';
-import { removeAnnotation } from '../features/annotation/annotationSlice';
+import { addAnnotation, clearSearches, addSearch } from 'redux/slices/annotationSlice';
+import { removeAnnotation } from '../redux/slices/annotationSlice';
 import TrackListener from 'components/tracks/TrackListener';
 import { resolveConfig } from 'prettier';
 // import './canola.gff'
